Binomial Refactoring to Handle Combinary Explosion
While implement a new example I need to refactor the Binomial to handle Combinatory
explosion.

1. New Example

**Product Manager approval of a new Subscription Funnel Optimization**

This is a new Example using binomial and Normal distributions.

This scenery show  a Product Manager analyzing the result of an A/B test for an
subscription funnel optimization approval using binomial and normal distribution
in order to have a solid understand of the results.

2. Refactoring to handle Big Numbers

or, the [Combinatory Explosion](https://pt.wikipedia.org/wiki/Explos%C3%A3o_combinat%C3%B3ria) Problem.

At First the Binomial implementation use Int64 as a conscious decision because
 I want to handle integer overflow but I knew that at some point a refactoring
would be necessary.

This is the moment.

In the middle of a use case implementation I need to use a k of 1.000 and it give me
the integer overflow.

I was prepared to move all code to math/big but I prompt with gemini (that I'm
using as pair) show me the possibility of use logarithms and indicate this as a
good practice in statistics because:

* Avoid ovewflow and underflow.
* Improve Numerical Stability.
* Simplifies Optimization.
* Extends the Factorial to Real Numbers.

From this I implement a private method [using Log](https://go.dev/src/math/lgamma.go)
and maintain the Coefficient public method with a minimal change on return type
to float64.

I also remote the overflow tests, change the expect value of tests to float64 and
change the comparision method because == is unreliable due the imprecision of floating-point arithmetic(IEEE 754) so I use a delta to compare.